Add skipped-turn event and Meiling ability; expand game phases and improve XML documentation
- Introduce new game state `SkipTurn` and event `OnTurnSkipped` to support abilities that bypass phases
- Refactor `GameCore.MainGame` to fully support phase-by-phase turn execution (`StartOfTurn`, `Incident`, `Draw`, `Main`, `Discard`, `EndOfTurn`, etc.)
- Implement Hong Meiling’s ability: player may skip their turn to draw 4 cards using `OnStartOfTurn` and `OnTurnSkipped` events
- Improve Meiling’s custom `ModifierData` to apply only in valid phases and to other players
- Add `Init()` method to `IExpansionData` for expansion-specific initialization during game startup
- Change `DefaultData` fields from `int` to `byte` to optimize memory usage
- Add full XML-style documentation to interfaces and models:
  - `IEventManager`, `BaseEventArgs`, `NamedObject`, `Durations`, `Seasons`, and other foundational classes now have complete `<summary>` and `<param>` tags

+++ DanmakuCardGameEngine/Core/DefaultData.cs
@@ -1,21 +1,21 @@
 ﻿namespace DanmakuCardGameEngine.Core {
     public sealed class DefaultData : IDefaultData {
-        public int DanmakuLimit => 1;
-        public int MaxLife => 4;
-        public int MaxHandSize => 4;
-        public int Range => 1;
-        public int Distance => 1;
-        public int CardDraw => 2;
+        public byte DanmakuLimit => 1;
+        public byte MaxLife => 4;
+        public byte MaxHandSize => 4;
+        public byte Range => 1;
+        public byte Distance => 1;
+        public byte CardDraw => 2;
 
         internal DefaultData() { }
     }
 
     public interface IDefaultData {
-        int DanmakuLimit { get; }
-        int MaxLife { get; }
-        int MaxHandSize { get; }
-        int Range { get; }
-        int Distance { get; }
-        int CardDraw { get; }
+        byte DanmakuLimit { get; }
+        byte MaxLife { get; }
+        byte MaxHandSize { get; }
+        byte Range { get; }
+        byte Distance { get; }
+        byte CardDraw { get; }
     }
 }
